input: move mouse handler code to cpp
