+++ rpcs3/emucore.vcxproj.filters
@@ -1252,6 +1252,7 @@
     <ClCompile Include="Emu\Io\mouse_config.cpp">
       <Filter>Emu\Io</Filter>
     </ClCompile>
+    <ClCompile Include="Emu\Io\MouseHandler.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Crypto\aes.h">
